Fixed deposit logic for EtherDelta
It's not perfect as it may not correctly in case of more than one band
on each side. But now at least it doesn't try to create orders larger
than the amount deposited to EtherDelta, which used to fail directly in
the API.

The edge case for multiple bands on one side is that if deposited amount
will be higher than the size of each individual band, but lower than the
sum of these two bands, the keeper will not deposit more even if it has
enough balance available.