Building an in-memory order book in order to reflect pending operations
With this approach, a newly placed order is present in the 'virtual'
order book the keeper uses to take decisions, which for example makes it
possible for the keeper to cancel that order before we even have a
chance to fully refresh the orderbook from the server.

At the same time, `in_progress` becomes False immediately after order
cancellation is successful, which means the keeper can replace the
freshly cancelled orders with a new one without having to wait for a
full orderbook refresh from the server.