Changed the way EtherDelta deposits work
Previously, we did deposit to EtherDelta only if the deposited amount
was lower than the value of one of the bands. So if we have two sell
bands of 20k each, and deposited amount is 30k, the keeper would not
deposit, as 30k > 20k.

In the new approach (with this fix), it will deposit as 30k < 20k+20k.

The old approach did work with EtherDelta as open orders do not decrease
available balance, but with other depositable exchanges like IDEX it
wouldn't work. That's why I want to standardize on this logic.

With this approach btw it will be much easier to reuse the new order
placement logic present in the `Bands` class.