Distringuished between total and available balances in method names
Some exchanges return available balance (for these `our_available_balance`
is implemented), and some total balance (for these `our_total_balance`
is impkemented) and total amount of all open orders shall be subtracted
from that total balance before placing new ones - so we do not try do
reuse the same balances twice. Altgough it would work in the case of
EtherDelta and RadarRelay, we try not to do it.