Performed an initial README bands update and received the following feedback from Lev: ``` I would change it so that the buy and sell examples use similar numbers, i.e. are both ETH/DAI rather than one being DAI/USD and the other being ETH/USD. That would probably be less confusing.
Also the sell example is wrong: if ETH drops from $100 to $75 (-25%) then all of the existing orders will be out of the range of the bands, since the size of the bands is only 1.5% and 1% respectively.
```
I adjusted the readme to compensate for the critique.

This PR adds examples to the bands file in the hopes of making the following aspects of the bands easier to understand: